+++ AndroidSrc/unitychirpio/src/main/java/com/alexanderkub/plugins/unitychirpio/ChirpPluginJavaMessageHandler.java
@@ -1,7 +1,7 @@
 package com.alexanderkub.plugins.unitychirpio;
 
 public interface ChirpPluginJavaMessageHandler {
-    void OnChangeStateHandler(byte state);
+    void OnChangeStateHandler(int state);
     void OnReceiveDataHandler(String message);
     void OnSentDataHandler(String message);
 }
